move the pdf views under common/views/pdf
