+++ pages/common/views/pdf/footer.jsx
