+++ pages/common/views/pdf/standard-conditions.jsx
