+++ pages/common/views/pdf/status-banner.jsx
